More sound work, basic retriggerable asset
Made the basic premise for a retriggerable sound that only plays once a previous iteration has stopped playing. It might be better if the hit sounds are fairly short so that they can be better activated. As it is the sound goes on for a fairly long time after the hit